+++ User/systick.c
@@ -1,6 +1,6 @@
 
 #include "systick.h"
 
-uint32_t systick_GetTick() {
+uint64_t systick_GetTick() {
     return systick_value;
 }
