+++ User/systick.h
@@ -1,6 +1,6 @@
 
 #include <stdint.h>
 
-extern uint32_t systick_value;
+extern uint64_t systick_value;
 
-uint32_t systick_GetTick(void);
+uint64_t systick_GetTick(void);
